% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppallytest-package.R
\docType{package}
\name{cppallytest-package}
\alias{cppallytest}
\alias{cppallytest-package}
\title{cppallytest: A Package For Testing cppally}
\description{
Unit tests for the R package cppally.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/cpp20/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
\keyword{internal}
